+++ src/components/buttons/AnswerButton/AnswerButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import cn from "classnames";
 
-import "./correct-button.scss";
+import "./answer-button.scss";
 
 export interface AnswerButtonProps {
   label?: string;
@@ -9,6 +9,7 @@ export interface AnswerButtonProps {
   className?: string;
   round?: boolean;
   onClick?: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
+  selected?: boolean;
 }
 
 const AnswerButton = ({
@@ -16,18 +17,20 @@ const AnswerButton = ({
   cancel,
   className,
   round,
-  onClick
+  onClick,
+  selected
 }: AnswerButtonProps) => {
   return (
     <button
       data-testid="answer-button"
-      className={cn("correct-button", className, {
-        ["correct-button_cancel"]: cancel,
-        ["correct-button_round"]: round
+      className={cn("answer-button", className, {
+        ["answer-button_cancel"]: cancel,
+        ["answer-button_round"]: round,
+        ["answer-button__selected"]: selected
       })}
       onClick={onClick}
     >
-      <div className="correct-button_image-container" />
+      <div className="answer-button_image-container" />
 
       <span>{!round && label}</span>
     </button>